Busqueda Combinada ahora busca en las referencias de nivel 3, se puede buscar por disponibilidad
git-svn-id: svn+ssh://163.10.10.34/var/repositorioV2-V3@2256 107f0ce8-b514-0410-b978-fdcfc4524ee3

+++ intranet/modules/C4/AR/Reservas.pm
@@ -910,13 +910,14 @@ sub manejoDeDisponibilidadDomiciliaria{
         #retorna las reserva en espera (SI EXISTEN) del grupo
         my ($reservas_en_espera_array_ref) = getReservasEnEsperaById2($db, $params->{'id2'}); 
 
-        foreach my $r (@$reservas_en_espera_array_ref) {
-            C4::AR::Debug::debug("manejoDeDisponibilidadDomiciliaria => elimino la reserva con id_reserva: ".$r->getId_reserva);
-        #elimino todas las sanciones y las reservas
-            $r->borrar_sancion_de_reserva($params->{'db'});
-            $r->delete();
+        if($reservas_en_espera_array_ref){
+          foreach my $r (@$reservas_en_espera_array_ref) {
+              C4::AR::Debug::debug("manejoDeDisponibilidadDomiciliaria => elimino la reserva con id_reserva: ".$r->getId_reserva);
+          #elimino todas las sanciones y las reservas
+              $r->borrar_sancion_de_reserva($params->{'db'});
+              $r->delete();
+          }
         }
-
     }
 }
